voicegpt/requirements.txt
requests
scikit-learn
pandas
flytekit
flytekitplugins-deck-standard
google-cloud-speech
google-cloud-texttospeech
marvin
pyaudio
sounddevice
numpy
scipy
asyncio